requirements.txt
selenium>=4.21.0
python-dotenv>=1.0.1